% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chp_6_Zero_tests.R
\name{k_At_Nconf}
\alias{k_At_Nconf}
\title{Characteristic Life Multiplier}
\usage{
k_At_Nconf(n, beta, conf = 0.9)
}
\arguments{
\item{n}{number of components tested simultaneously}

\item{beta}{shape of the Weibull distribution}

\item{conf}{confidence to demonstrate}
}
\description{
[ Section 6.1 New Weibull Handbook ] 
The characteristic life multiplier gives a constant that will indicate
for how much or less the test will have to be carry out, depending on the inputs. 
The results are based on the null hypothesis, that the new design is no better than the old.
}
\examples{
n <- 1:10
k_At_Nconf(n, beta= 1.5, conf=.9)
}

