% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chp_6_Zero_tests.R
\name{Probability_Passing_Zero_Test}
\alias{Probability_Passing_Zero_Test}
\title{Probability pass zero test}
\usage{
Probability_Passing_Zero_Test(n, beta, r, conf = 0.9)
}
\arguments{
\item{n}{number of components tested simultaneously}

\item{beta}{shape of the Weibull distribution}

\item{r}{ratio: true eta / demonstrated eta}

\item{conf}{confidence to demonstrate}
}
\description{
NEED TO UNDERSTAND THIS!
}
\examples{
Probability_Passing_Zero_Test(n=10, beta=1.5, r=.5, conf=.9)
}

