% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chp_6_Zero_tests.R
\name{time_Pass_Zero_failure}
\alias{time_Pass_Zero_failure}
\title{Zero test failure time}
\usage{
time_Pass_Zero_failure(n, beta, dem_eta, conf = 0.9)
}
\arguments{
\item{n}{number of components tested simultaneously}

\item{beta}{shape of the Weibull distribution}

\item{dem_eta}{Requirement to demonstrate new eta (scale)}

\item{conf}{confidence to demonstrate}
}
\description{
The zero test failure calculates how much time a test with zero failures 
have to run in order to demonstrate a certain eta with an specific level of confidence.
If the one or more of the components fail, the test has fail. Making several Zero tests
is consider cheating.
}
\examples{
time_Pass_Zero_failure(n=10, beta=1.5, dema_eta=500)

}

